feat: rename server utils module. resolve the module dependencies.
