+++ src/facefl/face_verification/simulation.py
@@ -22,7 +22,7 @@ from facefl.model.base_model import Net
 from facefl.model.driver import test, verify
 from facefl.server.wandb_server import WandbServer
 from facefl.simulation.app import start_simulation
-from facefl.utils.utils_server import load_strategy
+from facefl.server.strategy.common import load_strategy
 from facefl.utils.utils_wandb import custom_wandb_init
 
 warnings.filterwarnings("ignore")
